requirements.txt
requests
python-whois
termcolor
